refactor: separate useReadMore as custom hook
